// src/login-web-app/src/shared/util/types/media.types.ts
/*
 * Copyright (C) 2025 Curity AB. All rights reserved.
 *
 * The contents of this file are the property of Curity AB.
 * You may not copy or use this file, in either source code
 * or executable form, except in compliance with terms
 * set by Curity AB.
 *
 * For further information, please contact Curity AB.
 */

export enum MEDIA_TYPES {
  FORM_URLENCODED = 'application/x-www-form-urlencoded',
  JSON = 'application/json',
  AUTH = 'application/vnd.auth+json',
  PROBLEM = 'application/problem+json',
}